# storage/api-docs/api-docs.yaml
openapi: 3.0.0
info:
  title: 'Links API'
  description: ...
  version: 1.0.0
paths:
...
  /groups:
    get:
      tags:
        - Groups
      summary: 'Get paginated list of groups.'
      description: 'If admin, shows all existing ones. If a user, then only his groups.'
      operationId: cfc7c3d292050d13e480039da7bc441b
      responses:
        200:
          description: 'Paginated list.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GroupsListPaginated'
        401:
          description: Unauthenticated.
        403:
          description: 'Permissions error.'
      security:
        -
          sanctum: []
    post:
      tags:
        - Groups
      summary: 'Create new group.'
      operationId: 7822900eccb4f28c08960cd64607dc2b
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StoreGroupRequest'
      responses:
        200:
          description: 'Paginated list.'
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                    example: 'Info message.'
                  entity:
                    $ref: '#/components/schemas/Group'
                type: object
        401:
          description: Unauthenticated.
        403:
          description: 'Permissions error.'
        422:
          description: 'Validation error.'
      security:
        -
          sanctum: []
  '/groups/{id}':
    get:
      tags:
        - Groups
      summary: 'Show group by id.'
      operationId: 7450ba099c5d5127dbe6ceb30c5c8a44
      parameters:
        -
          name: id
          in: path
          description: 'ID of group.'
          required: true
          example: '1'
      responses:
        200:
          description: 'Group object.'
          content:
            application/json:
              schema:
                properties:
                  entity:
                    $ref: '#/components/schemas/Group'
                type: object
        401:
          description: Unauthenticated.
        403:
          description: 'Permissions error.'
        404:
          description: 'User not found.'
      security:
        -
          sanctum: []
    put:
      tags:
        - Groups
      summary: 'Update all group data.'
      operationId: 06accf2e97f47afe7eebb09f40112f84
      parameters:
        -
          name: id
          in: path
          description: 'ID of group.'
          required: true
          example: '1'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PutGroupRequest'
      responses:
        200:
          description: 'Group updated.'
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                    example: 'Info message.'
                  entity:
                    $ref: '#/components/schemas/Group'
                  old_entity:
                    $ref: '#/components/schemas/Group'
                type: object
        401:
          description: Unauthenticated.
        403:
          description: 'Permissions error.'
        404:
          description: 'Group not found.'
        422:
          description: 'Validation error.'
      security:
        -
          sanctum: []
    delete:
      tags:
        - Groups
      summary: 'Delete group.'
      operationId: 46134f605004e58a96c3da4ad69e41e1
      parameters:
        -
          name: id
          in: path
          description: 'ID of group.'
          required: true
          example: '1'
      responses:
        200:
          description: 'Group deleted.'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        401:
          description: Unauthenticated.
        403:
          description: 'Permissions error.'
        404:
          description: 'Group not found.'
      security:
        -
          sanctum: []
    patch:
      tags:
        - Groups
      summary: 'Partly update group data.'
      operationId: 580abc69157d81a2e9f7dea753d19e47
      parameters:
        -
          name: id
          in: path
          description: 'ID of group.'
          required: true
          example: '1'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchGroupRequest'
      responses:
        200:
          description: 'Group updated.'
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                    example: 'Info message.'
                  entity:
                    $ref: '#/components/schemas/Group'
                  old_entity:
                    $ref: '#/components/schemas/Group'
                type: object
        401:
          description: Unauthenticated.
        403:
          description: 'Permissions error.'
        404:
          description: 'Group not found.'
        422:
          description: 'Validation error.'
      security:
        -
          sanctum: []
...
components:
  schemas:
    MessageResponse:
      properties:
        message:
          type: string
          example: 'Info message'
      type: object
...
    PatchGroupRequest:
      properties:
        name:
          type: string
          example: 'Example group'
        description:
          type: string
          example: 'Example description'
        count:
          type: integer
          example: '10'
      type: object
    PutGroupRequest:
      required:
        - name
        - description
        - count
      properties:
        name:
          type: string
          example: 'Example group'
        description:
          type: string
          example: 'Example description'
        count:
          type: integer
          example: '10'
      type: object
...
    StoreGroupRequest:
      required:
        - name
        - description
        - origin
      properties:
        name:
          type: string
          example: 'Example group'
        description:
          type: string
          example: 'Example description'
        origin:
          type: string
          example: 'https://example.url/11111'
        group_id:
          type: array
          items:
            type: integer
            example: '1'
      type: object
    Group:
      properties:
        id:
          type: integer
          example: '1'
        name:
          type: string
          example: 'Example group'
        description:
          type: string
          example: 'Example description'
        count:
          type: integer
          example: '0'
        created_at:
          type: string
          format: date-time
          example: '2023-09-27T06:16:30.000000Z'
        updated_at:
          type: string
          format: date-time
          example: '2023-09-27T06:16:30.000000Z'
      type: object
    GroupsListPaginated:
      properties:
        current_page:
          type: integer
          example: '1'
        data:
          type: array
          items:
            $ref: '#/components/schemas/Group'
        first_page_url:
          type: string
          example: 'http://links.com/groups?page=1'
        from:
          type: integer
          example: '1'
        last_page:
          type: integer
          example: '2'
        last_page_url:
          type: string
          example: 'http://links.com/groups?page=2'
        links:
          type: array
          items:
            properties:
              url:
                type: string
                example: 'null | http://links.com/groups?page=2'
              label:
                type: string
                example: 'Next &raquo'
              active:
                type: boolean
                example: 'false'
            type: object
        next_page_url:
          type: string
          example: 'null | http://links.com/groups?page=2'
        path:
          type: string
          example: 'http://links.com/groups'
        per_page:
          type: integer
          example: '10'
        prev_page_url:
          type: string
          example: 'null | http://links.com/groups?page=2'
        to:
          type: integer
          example: '10'
        total:
          type: integer
          example: '15'
      type: object
...
  securitySchemes:
    sanctum:
      type: apiKey
      description: 'Enter token in format (Bearer <token>)'
      name: Authorization
      in: header